Move `pathutil` package into `cmdexport`
